+++ frontend/fourletterdisplay.config.js
@@ -11,64 +11,51 @@ function($rootScope, cleepService, fourletterdisplayService) {
         var self = this;
         self.config = {};
         self.message = '';
-        self.dots = [false, false, false, false];
+        self.dotsOptions = [
+            { label: 'left dot', value: 0 },
+            { label: 'middle-left dot', value: 1 },
+            { label: 'middle-right dot', value: 2 },
+            { label: 'right dot', value: 3 },
+        ];
+        self.selectedDots = [];
 
-        /**
-         * Display message
-         */
         self.displayMessage = function() {
             fourletterdisplayService.displayMessage(self.message);
+            self.message = '';
         };
 
-        /**
-         * Set dots
-         */
-        self.setDots = function() {
-            fourletterdisplayService.setDots(self.dots[0], self.dots[1], self.dots[2], self.dots[3]);
+        self.setDots = function(value) {
+            const dots = new Array(4).fill(false);
+            value.forEach((dot) => dots[dot] = true);
+            fourletterdisplayService.setDots(dots[0], dots[1], dots[2], dots[3]);
         };
 
-        /**
-         * Enable night mode
-         */
-        self.enableNightMode = function() {
-            fourletterdisplayService.enableNightMode(self.config.nightmode)
+        self.enableNightMode = function(value) {
+            fourletterdisplayService.enableNightMode(value)
                 .then(function(resp) {
                     cleepService.reloadModuleConfig('fourletterdisplay')
                 });
         };
 
-        /**
-         * Set default brightness
-         */
-        self.setBrightness = function() {
-            fourletterdisplayService.setBrightness(self.config.brightness)
+        self.setBrightness = function(value) {
+            fourletterdisplayService.setBrightness(value)
                 .then(function(resp) {
                     cleepService.reloadModuleConfig('fourletterdisplay')
                 });
         };
 
-        /**
-         * Set night mode brightness
-         */
-        self.setNightModeBrightness = function() {
-            fourletterdisplayService.setNightModeBrightness(self.config.nightbrightness)
+        self.setNightModeBrightness = function(value) {
+            fourletterdisplayService.setNightModeBrightness(value)
                 .then(function(resp) {
                     cleepService.reloadModuleConfig('fourletterdisplay');
                 });
         };
 
-        /**
-         * Clear display
-         */
-        self.clear = function() {
+        self.clearDisplay = function() {
             fourletterdisplayService.clear();
         };
 
-        /**
-         * Init component
-         */
         self.$onInit = function() {
-            // get module config
             cleepService.getModuleConfig('fourletterdisplay');
         };
 
@@ -89,7 +76,7 @@ function($rootScope, cleepService, fourletterdisplayService) {
         replace: true,
         scope: true,
         controller: fourletterdisplayConfigController,
-        controllerAs: 'fourletterdisplayCtl',
+        controllerAs: '$ctrl',
     };
 }]);