+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Fixed
 - Fix documentation
 
+### Updated
+- Use new config components from Cleep core
+
 ## [1.1.0] - 2023-03-13
 ### Changed
 - Update on_render parameter names according to core